% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsing.R
\name{bm_get_odds_xml}
\alias{bm_get_odds_xml}
\title{Download and read odds in xml format from bookmaker.eu into R}
\usage{
bm_get_odds_xml(url = "http://lines.bookmaker.eu")
}
\arguments{
\item{url}{domain or path of the xml file}
}
\value{
`xml_document`
}
\description{
Download and read odds in xml format from bookmaker.eu into R
}
\examples{
odds <- bm_get_odds_xml()

}
