% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{bm_mk_reactable_tbl}
\alias{bm_mk_reactable_tbl}
\title{Reactable defaults}
\usage{
bm_mk_reactable_tbl(df, grp_by)
}
\arguments{
\item{df}{dataframe}

\item{grp_by}{columns to group reactable by}
}
\value{
reactable table
}
\description{
Reactable defaults
}
